update table server example with the new workflow
