+++ packages/react-vapor/docs/Reducers.ts
@@ -4,10 +4,7 @@ import {
     IListBoxExampleCompositeState,
     listBoxExampleReducer,
 } from '../src/components/listBox/examples/ListBoxExampleReducer';
-import {
-    IExampleServerTableState,
-    TableHOCServerExampleReducer,
-} from '../src/components/table-hoc/examples/TableHOCServerExamples';
+import {IExampleServerTableState} from '../src/components/table-hoc/examples/TableHOCServerExample';
 import {IReactVaporState} from '../src/ReactVapor';
 import {ReactVaporReducers} from '../src/ReactVaporReducers';
 
@@ -19,5 +16,4 @@ export interface IReactVaporExampleState extends IReactVaporState {
 export const Reducers: Redux.Reducer<IReactVaporExampleState> = Redux.combineReducers<IReactVaporExampleState>({
     ...ReactVaporReducers,
     listBoxExampleState: listBoxExampleReducer,
-    tableHOCExample: TableHOCServerExampleReducer,
 });
